+++ CHANGES.md
@@ -8,6 +8,10 @@ Version 0.2.0
 
 To be released.
 
+ -  Added `MerkleHash#compareTo()` method.  [[#4]]
+
+[#4]: https://github.com/dahlia/json-hash/issues/4
+
 
 Version 0.1.0
 -------------